use set function instead of merge
merge does a deep merge, but that could cause weird problems
if two different livenessProbe definitions were merged and
other things that would not make sense.

Instead, we use set to mutate the $sensor dictionary applying
the override values on top of the defaults, effectively merging
only one level of the dictionaries.