Allow excluding default/sample RBAC files in values
As values dictionaries are merged, there is no way to remove
unwanted rbac files like sample.yaml. And, we cannot assume
that st2admin and stanley are present, because those are configurable.

So, we allow the user to set the sample/default files to an empty string
to prevent them from inclusion in their stackstorm-ha cluster.