Add pre-commit hook for common misspelling check in files (apache/airflow#18964)
This PR adds codespell to the pre-commit hooks. This will specifically help
us a bit in resolving sphinx errors.

From the project page:
It does not check for word membership in a complete dictionary, but instead looks for a set of common misspellings.
Therefore it should catch errors like "adn", but it will not catch "adnasdfasdf".
This also means it shouldn't generate false-positives when you use a niche term it doesn't know about.

This means the sphinx errors are not solved completely.

Partial Commit Extracted From: https://github.com/apache/airflow