Improve migration script Deployment handling
Helm could not apply some of the 1.0.0 changes to deployments because
some fields were immutable. So, this adjusts the migration script
to orphan the replicasets and delete the deployments. Then when helm upgrade
recreates the deployments, they will adopt the orphaned replicasets and
gradually recreate both replicasets and pods using the latest spec.