Allow clusters that do not use kubeproxy to disable presleep
When I tried using "0", it was somehow converted to an integer in the yaml file
causing some failures during helm upgrade. So, this also adds quotes around the
preStopSleep.

Also, the changelog was missing the PR that added the preStopSleep, so add it,
and fix the order of another changelog entry to follow the merge order and
link to the PR instead of the issue.