Allow clusters that do not use kubeproxy to disable presleep (#416)
Adds a conditional to skip adding the preStop sleep (added in #382) for clusters that
do not use kubelet. My cluster, for example, uses the Calico eBPF-based dataplane,
so it does not rely on kubelet or iptables.

When I tried using "0" to skip this, it was somehow converted to an integer in the
generated yaml file causing some failures during helm upgrade. So, this also adds
quotes around the preStopSleep.

Also, the changelog was missing the PR that added the preStopSleep, so add it, and
fix the order of another changelog entry to follow the merge order and link to the
PR instead of the issue.